pyproject.toml
[project]
name = "oism-note"
version = "0.1.0"
description = "Add a short description here"
authors = [
    { name = "sinyo-matu", email = "sinyo_matu@icloud.com" }
]
dependencies = [
    "ipykernel>=6.23.2",
    "pandas>=2.0.2",
    "ipython-sql>=0.5.0",
    "matplotlib>=3.7.1",
    "seaborn>=0.12.2",
    "polars>=0.18.3",
    "connectorx>=0.3.1",
    "pyarrow>=12.0.1",
    "plotly>=5.15.0",
    "nbformat>=5.9.0",
    "nbconvert>=7.5.0",
    "six>=1.16.0",
    "jupyter>=1.0.0",
    "pymongo>=4.7.2",
    "python-dotenv>=1.0.1",
    "pydantic>=2.7.1",
    "httpx>=0.27.0",
]
readme = "README.md"
requires-python = ">= 3.10"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true

[tool.hatch.metadata]
allow-direct-references = true
